requirements.txt
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.12.0
azure-keyvault-secrets==4.7.0
cachetools==5.3.0
camel-morph==0.9.1
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5d968ni_yn/croot/certifi_1671487774636/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
emoji==2.2.0
Flask==2.2.3
google-auth==2.17.2
google-auth-oauthlib==1.0.0
gspread==5.1.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.2.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msal==1.21.0
msal-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
pandas==2.0.0
portalocker==2.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
